Time: 0 ms (100.00%) | Memory: 8.2 MB (33.14%) - LeetSync

+++ 1078-remove-outermost-parentheses/remove-outermost-parentheses.cpp
@@ -2,11 +2,11 @@ class Solution {
 public:
     string removeOuterParentheses(string s) {
         stack<char> st;
-        string ans = "";
-        for(int i = 0; i < (int) s.size(); i++) {
-            if(s[i] == ')') st.pop();
-            if(!st.empty()) ans += s[i];
-            if(s[i] == '(') st.push(s[i]);
+        string ans ="";
+        for(int i=0;i<s.size();i++){
+            if(s[i]==')') st.pop();
+            if(!st.empty()) ans+=s[i];
+            if(s[i]=='(') st.push(s[i]);
         }
         return ans;
     }
